Increased click-sensitive area of accumulator to include "ACCUMULATOR".
The previous version only responded to clicks in the accumulator itself.
However, it is sometimes easier to describe clicking on the word than
on the box above the word. This version accepts both. Note, however that
there is a trade-off between larger clickable regions for controls like
the accumulator and encroachment on the clickable area for sliders that
pass under those areas. Until a natural alternative is found, these
"special" clicking regions should be kept to a minimum.