Made I/O Covers transparent in transparent mode (not as a slider)
The I/O Covers had been drawn as a slider (that had no room to slide)
because that code already existed. However, it made more sense to
create it as a static image just like the computer itself. This commit
makes that change and also adds a transparent version of the I/O Covers