Removed the need for transparent I/O Covers by rearranging stack
The I/O Covers were needed to cover the I/O sliders to make them
appear as if they went behind the computer card. This gives the
normal (topmost first) stacking order of:

  I/O Covers
  I/O Sliders
  Computer
  CPU Sliders

But since the I/O Covers are intentionally non-transparent, this
stacking order would require a transparent version for transparent
mode. However, the need for a transparent version can be eliminated
by realizing that the tranparent computer could go on top, and it
already has the "red ovals" for the I/O slider holes. So this gives
a transparent stacking order of:

  Computer (transparent)
  I/O Sliders
  CPU Sliders

In other words, the I/O covers aren't needed if there's nothing to
be covered in transparent mode. As a result, the previously added
transparent cover image (IOCOV_X.png) is no longer needed and has
been removed in this commit.